﻿[wip]: Add login
-- To Commit
Going to try to implement typegoose
this is because at the moment of declaring the model with ts-mongoose
and use the model in other folder the

let user = User.findOne()

the user does not infered the type. I need to specify it.

I was migrating to ts-mongoose because I through this will be resolved.
If this is the case I think typegoose is going to be the path to go.